refactor: artist 상세 페이지 컴포넌트 폴더명 _components로 통일
