+++ apps/client/app/[exhibitionId]/artist/[artistId]/_components/Navigation/Navigator.styles.ts
