+++ apps/client/app/[exhibitionId]/artist/[artistId]/_components/Navigation/PostNavigation/PostNavigation.tsx
