+++ apps/client/app/[exhibitionId]/artist/[artistId]/_components/sections/ArtworkSection.tsx
