+++ apps/client/app/[exhibitionId]/artist/[artistId]/_components/sections/BTSSection.tsx
