+++ apps/client/app/[exhibitionId]/artist/[artistId]/_components/sections/ProfileSection.tsx
